requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
astroid==3.3.11
attrs==25.3.0
audioop-lts==0.2.2
cffi==1.17.1
dill==0.4.0
discord.py==2.5.2
frozenlist==1.7.0
idna==3.10
iniconfig==2.1.0
isort==6.0.1
mccabe==0.7.0
multidict==6.6.3
packaging==25.0
platformdirs==4.3.8
pluggy==1.6.0
propcache==0.3.2
pycparser==2.22
Pygments==2.19.2
pylint==3.3.7
PyNaCl==1.5.0
pytest==8.4.1
python-dotenv==1.1.1
tomlkit==0.13.3
yarl==1.20.1
